@@ -12,10 +12,13 @@
         }
     ],
     "require": {
-        "php": "^5.4|^7.0"
+        "php": "^8.2"
+    },
+    "scripts": {
+        "phpstan": "vendor/bin/phpstan analyse -c phpstan.neon --memory-limit=-1 src/"
     },
     "require-dev": {
-        "phpunit/phpunit": "~4.0",
+        "phpunit/phpunit": "10.4.2",
         "mockery/mockery": "~0.8",
         "guzzlehttp/guzzle": "~5.0"
     },
