@@ -24,8 +24,10 @@
 namespace Facebook\Tests\Authentication;
 
 use Facebook\Authentication\AccessTokenMetadata;
+use Facebook\Exceptions\FacebookSDKException;
+use PHPUnit\Framework\TestCase;
 
-class AccessTokenMetadataTest extends \PHPUnit_Framework_TestCase
+class AccessTokenMetadataTest extends TestCase
 {
 
     protected $graphResponseData = [
@@ -81,57 +83,52 @@ class AccessTokenMetadataTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('1337', $metadata->getUserId());
     }
 
-    /**
-     * @expectedException \Facebook\Exceptions\FacebookSDKException
-     */
-    public function testInvalidMetadataWillThrow()
+	public function testInvalidMetadataWillThrow()
     {
-        new AccessTokenMetadata(['foo' => 'bar']);
+		$this->expectException(FacebookSDKException::class);
+		new AccessTokenMetadata(['foo' => 'bar']);
     }
 
     public function testAnExpectedAppIdWillNotThrow()
     {
+		$this->expectNotToPerformAssertions();
         $metadata = new AccessTokenMetadata($this->graphResponseData);
         $metadata->validateAppId('123');
     }
 
-    /**
-     * @expectedException \Facebook\Exceptions\FacebookSDKException
-     */
-    public function testAnUnexpectedAppIdWillThrow()
+	public function testAnUnexpectedAppIdWillThrow()
     {
-        $metadata = new AccessTokenMetadata($this->graphResponseData);
+		$this->expectException(FacebookSDKException::class);
+		$metadata = new AccessTokenMetadata($this->graphResponseData);
         $metadata->validateAppId('foo');
     }
 
     public function testAnExpectedUserIdWillNotThrow()
     {
+		$this->expectNotToPerformAssertions();
         $metadata = new AccessTokenMetadata($this->graphResponseData);
         $metadata->validateUserId('1337');
     }
 
-    /**
-     * @expectedException \Facebook\Exceptions\FacebookSDKException
-     */
-    public function testAnUnexpectedUserIdWillThrow()
+	public function testAnUnexpectedUserIdWillThrow()
     {
-        $metadata = new AccessTokenMetadata($this->graphResponseData);
+		$this->expectException(FacebookSDKException::class);
+		$metadata = new AccessTokenMetadata($this->graphResponseData);
         $metadata->validateUserId('foo');
     }
 
     public function testAnActiveAccessTokenWillNotThrow()
     {
+		$this->expectNotToPerformAssertions();
         $this->graphResponseData['data']['expires_at'] = time() + 1000;
         $metadata = new AccessTokenMetadata($this->graphResponseData);
         $metadata->validateExpiration();
     }
 
-    /**
-     * @expectedException \Facebook\Exceptions\FacebookSDKException
-     */
-    public function testAnExpiredAccessTokenWillThrow()
+	public function testAnExpiredAccessTokenWillThrow()
     {
-        $this->graphResponseData['data']['expires_at'] = time() - 1000;
+		$this->expectException(FacebookSDKException::class);
+		$this->graphResponseData['data']['expires_at'] = time() - 1000;
         $metadata = new AccessTokenMetadata($this->graphResponseData);
         $metadata->validateExpiration();
     }
