@@ -23,11 +23,12 @@
  */
 namespace Facebook\Tests\Authentication;
 
+use Facebook\Authentication\OAuth2Client;
 use Facebook\Facebook;
 use Facebook\FacebookApp;
-use Facebook\Authentication\OAuth2Client;
+use PHPUnit\Framework\TestCase;
 
-class OAuth2ClientTest extends \PHPUnit_Framework_TestCase
+class OAuth2ClientTest extends TestCase
 {
 
     /**
@@ -45,8 +46,9 @@ class OAuth2ClientTest extends \PHPUnit_Framework_TestCase
      */
     protected $oauth;
 
-    protected function setUp()
-    {
+    protected function setUp(): void
+	{
+		parent::setUp();
         $app = new FacebookApp('123', 'foo_secret');
         $this->client = new FooFacebookClientForOAuth2Test();
         $this->oauth = new OAuth2Client($app, $this->client, static::TESTING_GRAPH_VERSION);
@@ -79,7 +81,7 @@ class OAuth2ClientTest extends \PHPUnit_Framework_TestCase
         $scope = ['email', 'base_foo'];
         $authUrl = $this->oauth->getAuthorizationUrl('https://foo.bar', 'foo_state', $scope, ['foo' => 'bar'], '*');
 
-        $this->assertContains('*', $authUrl);
+        $this->assertStringContainsString('*', $authUrl);
 
         $expectedUrl = 'https://www.facebook.com/' . static::TESTING_GRAPH_VERSION . '/dialog/oauth?';
         $this->assertTrue(strpos($authUrl, $expectedUrl) === 0, 'Unexpected base authorization URL returned from getAuthorizationUrl().');
@@ -93,7 +95,7 @@ class OAuth2ClientTest extends \PHPUnit_Framework_TestCase
             'foo' => 'bar',
         ];
         foreach ($params as $key => $value) {
-            $this->assertContains($key . '=' . urlencode($value), $authUrl);
+            $this->assertStringContainsString($key . '=' . urlencode($value), $authUrl);
         }
     }
 
