@@ -27,8 +27,9 @@ use Facebook\FacebookApp;
 use Facebook\FacebookRequest;
 use Facebook\FacebookResponse;
 use Facebook\Exceptions\FacebookResponseException;
+use PHPUnit\Framework\TestCase;
 
-class FacebookResponseExceptionTest extends \PHPUnit_Framework_TestCase
+class FacebookResponseExceptionTest extends TestCase
 {
 
     /**
@@ -36,8 +37,9 @@ class FacebookResponseExceptionTest extends \PHPUnit_Framework_TestCase
      */
     protected $request;
 
-    protected function setUp()
+    protected function setUp(): void
     {
+		parent::setUp();
         $this->request = new FacebookRequest(new FacebookApp('123', 'foo'));
     }
 
