@@ -23,20 +23,22 @@
  */
 namespace Facebook\Tests;
 
+use Facebook\Exceptions\FacebookSDKException;
 use Facebook\Facebook;
 use Facebook\FacebookApp;
 use Facebook\FacebookRequest;
 use Facebook\FacebookBatchRequest;
 use Facebook\FileUpload\FacebookFile;
+use PHPUnit\Framework\TestCase;
 
-class FacebookBatchRequestTest extends \PHPUnit_Framework_TestCase
+class FacebookBatchRequestTest extends TestCase
 {
     /**
      * @var FacebookApp
      */
     private $app;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $this->app = new FacebookApp('123', 'foo_secret');
     }
@@ -80,32 +82,26 @@ class FacebookBatchRequestTest extends \PHPUnit_Framework_TestCase
         $this->assertRequestContainsAppAndToken($request, $customApp, 'foo_token');
     }
 
-    /**
-     * @expectedException \Facebook\Exceptions\FacebookSDKException
-     */
-    public function testWillThrowWhenNoThereIsNoAppFallback()
+	public function testWillThrowWhenNoThereIsNoAppFallback()
     {
-        $batchRequest = new FacebookBatchRequest();
+		$this->expectException(FacebookSDKException::class);
+		$batchRequest = new FacebookBatchRequest();
 
         $batchRequest->addFallbackDefaults(new FacebookRequest(null, 'foo_token'));
     }
 
-    /**
-     * @expectedException \Facebook\Exceptions\FacebookSDKException
-     */
-    public function testWillThrowWhenNoThereIsNoAccessTokenFallback()
+	public function testWillThrowWhenNoThereIsNoAccessTokenFallback()
     {
-        $request = new FacebookBatchRequest();
+		$this->expectException(FacebookSDKException::class);
+		$request = new FacebookBatchRequest();
 
         $request->addFallbackDefaults(new FacebookRequest($this->app));
     }
 
-    /**
-     * @expectedException \InvalidArgumentException
-     */
-    public function testAnInvalidTypeGivenToAddWillThrow()
+	public function testAnInvalidTypeGivenToAddWillThrow()
     {
-        $request = new FacebookBatchRequest();
+		$this->expectException(\InvalidArgumentException::class);
+		$request = new FacebookBatchRequest();
 
         $request->add('foo');
     }
@@ -168,22 +164,18 @@ class FacebookBatchRequestTest extends \PHPUnit_Framework_TestCase
         $this->assertRequestsMatch($requests, $formattedRequests);
     }
 
-    /**
-     * @expectedException \Facebook\Exceptions\FacebookSDKException
-     */
-    public function testAZeroRequestCountWithThrow()
+	public function testAZeroRequestCountWithThrow()
     {
-        $batchRequest = new FacebookBatchRequest($this->app, [], 'foo_token');
+		$this->expectException(FacebookSDKException::class);
+		$batchRequest = new FacebookBatchRequest($this->app, [], 'foo_token');
 
         $batchRequest->validateBatchRequestCount();
     }
 
-    /**
-     * @expectedException \Facebook\Exceptions\FacebookSDKException
-     */
-    public function testMoreThanFiftyRequestsWillThrow()
+	public function testMoreThanFiftyRequestsWillThrow()
     {
-        $batchRequest = $this->createBatchRequest();
+		$this->expectException(FacebookSDKException::class);
+		$batchRequest = $this->createBatchRequest();
 
         $this->createAndAppendRequestsTo($batchRequest, 51);
 
@@ -192,6 +184,8 @@ class FacebookBatchRequestTest extends \PHPUnit_Framework_TestCase
 
     public function testLessOrEqualThanFiftyRequestsWillNotThrow()
     {
+		$this->expectNotToPerformAssertions();
+
         $batchRequest = $this->createBatchRequest();
 
         $this->createAndAppendRequestsTo($batchRequest, 50);
@@ -213,9 +207,9 @@ class FacebookBatchRequestTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals($expectedArray, $batchRequestArray);
     }
 
-    public function requestsAndExpectedResponsesProvider()
+    public static function requestsAndExpectedResponsesProvider()
     {
-        $headers = $this->defaultHeaders();
+        $headers = self::defaultHeaders();
         $apiVersion = Facebook::DEFAULT_GRAPH_VERSION;
 
         return [
@@ -386,7 +380,7 @@ class FacebookBatchRequestTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals($expectedToken, $token);
     }
 
-    private function defaultHeaders()
+    private static function defaultHeaders()
     {
         $headers = [];
         foreach (FacebookRequest::getDefaultHeaders() as $name => $value) {
