@@ -28,8 +28,9 @@ use Facebook\FacebookRequest;
 use Facebook\FacebookResponse;
 use Facebook\FacebookBatchRequest;
 use Facebook\FacebookBatchResponse;
+use PHPUnit\Framework\TestCase;
 
-class FacebookBatchResponseTest extends \PHPUnit_Framework_TestCase
+class FacebookBatchResponseTest extends TestCase
 {
     /**
      * @var \Facebook\FacebookApp
@@ -41,7 +42,7 @@ class FacebookBatchResponseTest extends \PHPUnit_Framework_TestCase
      */
     protected $request;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $this->app = new FacebookApp('123', 'foo_secret');
         $this->request = new FacebookRequest(
