@@ -26,15 +26,16 @@ namespace Facebook\Tests;
 use Facebook\FacebookApp;
 use Facebook\FacebookRequest;
 use Facebook\FacebookResponse;
+use PHPUnit\Framework\TestCase;
 
-class FacebookResponseTest extends \PHPUnit_Framework_TestCase
+class FacebookResponseTest extends TestCase
 {
     /**
      * @var \Facebook\FacebookRequest
      */
     protected $request;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $app = new FacebookApp('123', 'foo_secret');
         $this->request = new FacebookRequest(
