@@ -23,6 +23,7 @@
  */
 namespace Facebook\Tests;
 
+use Facebook\Exceptions\FacebookSDKException;
 use Facebook\Facebook;
 use Facebook\FacebookClient;
 use Facebook\FacebookRequest;
@@ -33,20 +34,19 @@ use Facebook\Tests\Fixtures\FooBarPseudoRandomStringGenerator;
 use Facebook\Tests\Fixtures\FooClientInterface;
 use Facebook\Tests\Fixtures\FooPersistentDataInterface;
 use Facebook\Tests\Fixtures\FooUrlDetectionInterface;
+use PHPUnit\Framework\TestCase;
 
-class FacebookTest extends \PHPUnit_Framework_TestCase
+class FacebookTest extends TestCase
 {
     protected $config = [
         'app_id' => '1337',
         'app_secret' => 'foo_secret',
     ];
 
-    /**
-     * @expectedException \Facebook\Exceptions\FacebookSDKException
-     */
-    public function testInstantiatingWithoutAppIdThrows()
+	public function testInstantiatingWithoutAppIdThrows()
     {
-        // unset value so there is no fallback to test expected Exception
+		$this->expectException(FacebookSDKException::class);
+		// unset value so there is no fallback to test expected Exception
         putenv(Facebook::APP_ID_ENV_NAME.'=');
         $config = [
             'app_secret' => 'foo_secret',
@@ -54,12 +54,10 @@ class FacebookTest extends \PHPUnit_Framework_TestCase
         new Facebook($config);
     }
 
-    /**
-     * @expectedException \Facebook\Exceptions\FacebookSDKException
-     */
-    public function testInstantiatingWithoutAppSecretThrows()
+	public function testInstantiatingWithoutAppSecretThrows()
     {
-        // unset value so there is no fallback to test expected Exception
+		$this->expectException(FacebookSDKException::class);
+		// unset value so there is no fallback to test expected Exception
         putenv(Facebook::APP_SECRET_ENV_NAME.'=');
         $config = [
             'app_id' => 'foo_id',
@@ -67,12 +65,10 @@ class FacebookTest extends \PHPUnit_Framework_TestCase
         new Facebook($config);
     }
 
-    /**
-     * @expectedException \InvalidArgumentException
-     */
-    public function testSettingAnInvalidHttpClientHandlerThrows()
+	public function testSettingAnInvalidHttpClientHandlerThrows()
     {
-        $config = array_merge($this->config, [
+		$this->expectException(\InvalidArgumentException::class);
+		$config = array_merge($this->config, [
             'http_client_handler' => 'foo_handler',
         ]);
         new Facebook($config);
@@ -117,12 +113,10 @@ class FacebookTest extends \PHPUnit_Framework_TestCase
         );
     }
 
-    /**
-     * @expectedException \InvalidArgumentException
-     */
-    public function testSettingAnInvalidPersistentDataHandlerThrows()
+	public function testSettingAnInvalidPersistentDataHandlerThrows()
     {
-        $config = array_merge($this->config, [
+		$this->expectException(\InvalidArgumentException::class);
+		$config = array_merge($this->config, [
             'persistent_data_handler' => 'foo_handler',
         ]);
         new Facebook($config);
@@ -146,7 +140,7 @@ class FacebookTest extends \PHPUnit_Framework_TestCase
             ? 'TypeError'
             : 'PHPUnit_Framework_Error';
 
-        $this->setExpectedException($expectedException);
+        $this->expectException($expectedException);
 
         $config = array_merge($this->config, [
             'url_detection_handler' => 'foo_handler',
@@ -180,12 +174,10 @@ class FacebookTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('bar_token', (string)$accessToken);
     }
 
-    /**
-     * @expectedException \InvalidArgumentException
-     */
-    public function testSettingAnInvalidPseudoRandomStringGeneratorThrows()
+	public function testSettingAnInvalidPseudoRandomStringGeneratorThrows()
     {
-        $config = array_merge($this->config, [
+		$this->expectException(\InvalidArgumentException::class);
+		$config = array_merge($this->config, [
             'pseudo_random_string_generator' => 'foo_generator',
         ]);
         new Facebook($config);
@@ -273,12 +265,10 @@ class FacebookTest extends \PHPUnit_Framework_TestCase
         );
     }
 
-    /**
-     * @expectedException \InvalidArgumentException
-     */
-    public function testSettingAnAccessThatIsNotStringOrAccessTokenThrows()
+	public function testSettingAnAccessThatIsNotStringOrAccessTokenThrows()
     {
-        $config = array_merge($this->config, [
+		$this->expectException(\InvalidArgumentException::class);
+		$config = array_merge($this->config, [
             'default_access_token' => 123,
         ]);
         new Facebook($config);
@@ -402,12 +392,10 @@ class FacebookTest extends \PHPUnit_Framework_TestCase
         ], $response);
     }
 
-    /**
-     * @expectedException \Facebook\Exceptions\FacebookResponseException
-     */
-    public function testMaxingOutRetriesWillThrow()
+	public function testMaxingOutRetriesWillThrow()
     {
-        $client = new FakeGraphApiForResumableUpload();
+		$this->expectException(\Facebook\Exceptions\FacebookResponseException::class);
+		$client = new FakeGraphApiForResumableUpload();
         $client->failOnTransfer();
 
         $config = array_merge($this->config, [
