@@ -24,8 +24,9 @@
 namespace Facebook\Tests\FileUpload;
 
 use Facebook\FileUpload\Mimetypes;
+use PHPUnit\Framework\TestCase;
 
-class MimetypesTest extends \PHPUnit_Framework_TestCase
+class MimetypesTest extends TestCase
 {
 
     /**
