@@ -25,15 +25,16 @@ namespace Facebook\Tests\GraphNodes;
 
 use Mockery as m;
 use Facebook\GraphNodes\GraphNodeFactory;
+use PHPUnit\Framework\TestCase;
 
-abstract class AbstractGraphNode extends \PHPUnit_Framework_TestCase
+abstract class AbstractGraphNode extends TestCase
 {
     /**
      * @var \Facebook\FacebookResponse|\Mockery\MockInterface
      */
     protected $responseMock;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         parent::setUp();
         $this->responseMock = m::mock('\Facebook\FacebookResponse');
