@@ -23,12 +23,14 @@
  */
 namespace Facebook\Tests\GraphNodes;
 
+use Facebook\Exceptions\FacebookSDKException;
 use Facebook\FacebookApp;
 use Facebook\FacebookRequest;
 use Facebook\GraphNodes\GraphEdge;
 use Facebook\GraphNodes\GraphNode;
+use PHPUnit\Framework\TestCase;
 
-class GraphEdgeTest extends \PHPUnit_Framework_TestCase
+class GraphEdgeTest extends TestCase
 {
 
     /**
@@ -41,7 +43,7 @@ class GraphEdgeTest extends \PHPUnit_Framework_TestCase
         'previous' => 'https://graph.facebook.com/v7.12/998899/photos?pretty=0&limit=25&before=foo_before_cursor',
     ];
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $app = new FacebookApp('123', 'foo_app_secret');
         $this->request = new FacebookRequest(
@@ -55,12 +57,10 @@ class GraphEdgeTest extends \PHPUnit_Framework_TestCase
         );
     }
 
-    /**
-     * @expectedException \Facebook\Exceptions\FacebookSDKException
-     */
-    public function testNonGetRequestsWillThrow()
+	public function testNonGetRequestsWillThrow()
     {
-        $this->request->setMethod('POST');
+		$this->expectException(FacebookSDKException::class);
+		$this->request->setMethod('POST');
         $graphEdge = new GraphEdge($this->request);
         $graphEdge->validateForPagination();
     }
