@@ -26,8 +26,9 @@ namespace Facebook\Tests\Helpers;
 use Facebook\FacebookApp;
 use Facebook\FacebookClient;
 use Facebook\Helpers\FacebookPageTabHelper;
+use PHPUnit\Framework\TestCase;
 
-class FacebookPageTabHelperTest extends \PHPUnit_Framework_TestCase
+class FacebookPageTabHelperTest extends TestCase
 {
     protected $rawSignedRequestAuthorized = '6Hi26ECjkj347belC0O8b8H5lwiIz5eA6V9VVjTg-HU=.eyJhbGdvcml0aG0iOiJITUFDLVNIQTI1NiIsImlzc3VlZF9hdCI6MzIxLCJvYXV0aF90b2tlbiI6ImZvb190b2tlbiIsInVzZXJfaWQiOiIxMjMiLCJwYWdlIjp7ImlkIjoiNDIiLCJsaWtlZCI6dHJ1ZSwiYWRtaW4iOmZhbHNlfX0=';
 
