@@ -24,8 +24,9 @@
 namespace Facebook\Tests\Http;
 
 use Facebook\Http\RequestBodyUrlEncoded;
+use PHPUnit\Framework\TestCase;
 
-class RequestUrlEncodedTest extends \PHPUnit_Framework_TestCase
+class RequestUrlEncodedTest extends TestCase
 {
     public function testCanProperlyEncodeAnArrayOfParams()
     {
