@@ -23,6 +23,7 @@
  */
 namespace Facebook\Tests\HttpClients;
 
+use Facebook\Exceptions\FacebookSDKException;
 use Mockery as m;
 use Facebook\HttpClients\FacebookStreamHttpClient;
 
@@ -38,7 +39,7 @@ class FacebookStreamHttpClientTest extends AbstractTestHttpClient
      */
     protected $streamClient;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $this->streamMock = m::mock('Facebook\HttpClients\FacebookStream');
         $this->streamClient = new FacebookStreamHttpClient($this->streamMock);
@@ -110,12 +111,10 @@ class FacebookStreamHttpClientTest extends AbstractTestHttpClient
         $this->assertEquals(200, $response->getHttpResponseCode());
     }
 
-    /**
-     * @expectedException \Facebook\Exceptions\FacebookSDKException
-     */
-    public function testThrowsExceptionOnClientError()
+	public function testThrowsExceptionOnClientError()
     {
-        $this->streamMock
+		$this->expectException(FacebookSDKException::class);
+		$this->streamMock
             ->shouldReceive('streamContextCreate')
             ->once()
             ->andReturn(null);
