@@ -28,9 +28,10 @@ use Facebook\HttpClients\FacebookGuzzleHttpClient;
 use Facebook\HttpClients\FacebookStreamHttpClient;
 use Facebook\HttpClients\HttpClientsFactory;
 use GuzzleHttp\Client;
+use PHPUnit\Framework\TestCase;
 use PHPUnit_Framework_TestCase;
 
-class HttpClientsFactoryTest extends PHPUnit_Framework_TestCase
+class HttpClientsFactoryTest extends TestCase
 {
     const COMMON_NAMESPACE = 'Facebook\HttpClients\\';
     const COMMON_INTERFACE = 'Facebook\HttpClients\FacebookHttpClientInterface';
@@ -52,7 +53,7 @@ class HttpClientsFactoryTest extends PHPUnit_Framework_TestCase
     /**
      * @return array
      */
-    public function httpClientsProvider()
+    public static function httpClientsProvider()
     {
         $clients = [
           ['guzzle', self::COMMON_NAMESPACE . 'FacebookGuzzleHttpClient'],
