@@ -24,9 +24,10 @@
 namespace Facebook\Tests\PseudoRandomString;
 
 use Facebook\PseudoRandomString\PseudoRandomStringGeneratorFactory;
+use PHPUnit\Framework\TestCase;
 use PHPUnit_Framework_TestCase;
 
-class PseudoRandomStringFactoryTest extends PHPUnit_Framework_TestCase
+class PseudoRandomStringFactoryTest extends TestCase
 {
     const COMMON_NAMESPACE = 'Facebook\PseudoRandomString\\';
     const COMMON_INTERFACE = 'Facebook\PseudoRandomString\PseudoRandomStringGeneratorInterface';
@@ -48,7 +49,7 @@ class PseudoRandomStringFactoryTest extends PHPUnit_Framework_TestCase
     /**
      * @return array
      */
-    public function csprngProvider()
+    public static function csprngProvider()
     {
         $providers = [
           [null, self::COMMON_INTERFACE],
