@@ -24,24 +24,22 @@
 namespace Facebook\Tests\PseudoRandomString;
 
 use Facebook\Tests\Fixtures\MyFooBarPseudoRandomStringGenerator;
+use InvalidArgumentException;
+use PHPUnit\Framework\TestCase;
 
-class PseudoRandomStringGeneratorTraitTest extends \PHPUnit_Framework_TestCase
+class PseudoRandomStringGeneratorTraitTest extends TestCase
 {
-    /**
-     * @expectedException \InvalidArgumentException
-     */
-    public function testAnInvalidLengthWillThrow()
+	public function testAnInvalidLengthWillThrow()
     {
-        $prsg = new MyFooBarPseudoRandomStringGenerator();
+		$this->expectException(InvalidArgumentException::class);
+		$prsg = new MyFooBarPseudoRandomStringGenerator();
         $prsg->validateLength('foo_len');
     }
 
-    /**
-     * @expectedException \InvalidArgumentException
-     */
-    public function testALengthThatIsNotAtLeastOneCharacterWillThrow()
+	public function testALengthThatIsNotAtLeastOneCharacterWillThrow()
     {
-        $prsg = new MyFooBarPseudoRandomStringGenerator();
+		$this->expectException(InvalidArgumentException::class);
+		$prsg = new MyFooBarPseudoRandomStringGenerator();
         $prsg->validateLength(0);
     }
 }
