@@ -24,8 +24,9 @@
 namespace Facebook\Tests\Url;
 
 use Facebook\Url\FacebookUrlManipulator;
+use PHPUnit\Framework\TestCase;
 
-class FacebookUrlManipulatorTest extends \PHPUnit_Framework_TestCase
+class FacebookUrlManipulatorTest extends TestCase
 {
     /**
      * @dataProvider provideUris
@@ -44,7 +45,7 @@ class FacebookUrlManipulatorTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals($expectedCleanUrl, $currentUri);
     }
 
-    public function provideUris()
+    public static function provideUris()
     {
         return [
             [
@@ -162,7 +163,7 @@ class FacebookUrlManipulatorTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals($result, $expected);
     }
 
-    public function provideMergableEndpoints()
+    public static function provideMergableEndpoints()
     {
         return [
             [
